Functional options for steps (#59)
Our current method for capturing the reflection name of a step is
flawed. The typed erase function (anonymous) always has the same name,
and thus every call to RunAsStep overwrites it. This of course results
in wrong data in the `operation_results` table.

There is no good way to associated a unique key that the interface
method can use on this map. Hence, let's just move to the cleaner
functional options interface to handle step options, voiding the need
for the in-memory map.

Added a test (`TestSteps/checkStepName`) that fails without this change.