Conductor (#73)
Implement Conductor client.

# Design

A conductor object is attached to a `DBOSContext`. Its lifecycle
management is done like other resources, inside `DBOSContext` `Launch`
and `Shutdown` methods.

The conductor operates through two goroutines a "main" goroutine
handling conductor messages, and a "ping" goroutine to keep the
connection alive.

# Main goroutine

- Handles connecting to conductor
- Handles responding to conductor message
- Will stop only if its context is cancelled (root DBOSContext)

The main goroutine is entirely responsible for managing the connection:
whenever it detects it should connect to the server, it'll start a ping
goroutine. Whenever it gets an error (e.g., reading on the websocket or
sending a message), it will set the connection for closing.

# Ping goroutine

- Periodically sends keepalive pings to the server
- Handles pong answers

The ping goutine will exit and signal the connection should be close if:
- Its context is cancelled or
- It gets an error trying to send the ping

# Others

- There cannot be concurrent writers on the websocket, so all write
operations are protected by a mutex
- Shutdown logic: we cancel the ping goroutine, close the connection and
wait for both goroutine to terminate.
- Add a missing executors ID filter on `ListWorkflows`